Final Fixes 2
Fixed an issue where the main menu and quit game buttons in the game over screen in the Prototype_1_Level, Prototype_2_Level and Prototype_3_Level scenes were not interactable.

Fixed an index out of range error in the P2_InputManager.

Fixed a few bugs and issues related to when a player character dies and the game over event in the Prototype_2_Level scene.