Prototype 3 Improvements
Refined the reduction of the range and intensity of the lighthouse lights so that those values reduce at the same rate as the battery discharge rate.

Refined the difficulty settings for prototype 3, so that more enemies spawn with the increase in difficulty.

Implemented a new HUD element to remind the player of the input key to change characters.