Prototype 1 Improvements
Made small adjustments to the level design to reduce flickering on the ramps (due to hidden faces on the geometry) and to reduce the chance for the player to get stuck in the tutorial corridor.

Increased the sight and attack range values for all enemies. Also modified the walk point range variable depending on the type of enemy.

Slightly decreased the gun recoil move amount variable.