More Fixes and Improvements
Enabled the reload prompt in the P2_GameManager script so it is enabled when the scene loads.

Created an empty parent game object for the two player character selection HUD elements in the Prototype_2_Level so that the text moves when the primary visuals are disabled from the settings.

Attempted to fix the bug where the player can randomly shoot at a wall and randomly hit or kill an enemy that is elsewhere in the Prototype_2_Level scene.

Fixed a bug in the third prototype where the cursor wouldn't show after the scene restarted.

Updated the InputManager scripts for prototype 1 and 2 to reflect the changes made to the P3_InputManager script.